Making with_label accept an object responding to match
The current implementation of `with_label` accepts only a String
or a Regexp to be passed to match a light's label.

This commit simplifies the code and in the same time provides the
ability to give any object responding to `match` as a label.

String and Regexp both implement the `match` method, the original
behavior still works but it opens the door to other possibilities.